Add setvar option to calendaring
Adding the setvar option with variable substitution on the value allows things
like setting the outbound caller id name to the summary of a calendar event,
etc. Values could be chained together as they are appended in order to do some
scripting if necessary.

Review: https://reviewboard.asterisk.org/r/1134/