ConfBridge translation optimization
If multiple channels are using the same format, ConfBridge
does the encoding of the write frame and sends a single copy
out to all of them.  This drastically improves performance as
it requires only a single translation to take place per format.

Note, at the moment the 'drop_silence' option must be enabled to
take advantage of this feature.  This is because the optimization
can only take place when silence is detected on the channel in the
read direction. Otherwise the channel's own read audio has to be
mixed out of the conference mix which results in a unique frame for
that channel on write.

Also, some might consider this optimization of F7esh.