We do jitter buffering weird.  This branch is for experimenting on ways to address this.
Right now our jitter buffer only works on the write side of channels and is completely
bound to the generic bridge function.  This makes doing jitter buffering on channels
in conferences and other applications that are not bridged using this method impossible.
In order to get the best use out of a jitter buffer it should exist on the read side
of a channel.