The AsyncAGI command loop is lax in the value it returns for the return status.
* Return correct status: SUCCESS/FAILED/HANGUP.  Previously, abnormal
exits from the command loop such as hangup would return SUCCESS.

* The "asyncagi break" command now returns SUCCESS and is now the only way
to break the command loop with that status.  Previously, it returned
FAILED.

* The AMI event AsyncAGI End is no longer sent if the AsyncAGI Start event
is not sent.  Previously, this happened because of an error setting up the
AGI pipes.

* All executed AGI commands now get an AsyncAGI Exec result event.
Previously, if the command returned failure (because of hangup), the
command loop just exited with FAILURE and did not send the AsyncAGI Exec
result event.

* Makes sure that the channel frame queue is empty on hangup.

Review: https://reviewboard.asterisk.org/r/1183/