Branch for not breaking binary message bodies in chan_sip
SIP messages have a Content-Type header that could specify that the body is
binary data. Currently the message bodies are stored in an ast_str structure
and manipulated in ways that would break binary message bodies.

This branch is my playground for trying to come up with a way to fix this
limitation without breaking the world.