Handle the locking ourselves so we can chunk writes
This update removes all of the lovely threadstorage code so that we can easily
chunk writes to disk in 1 second bursts by wrapping statments in transactions.
Since we are handling the serialization ourselves, we now open the database in
single threaded mode so SQLite will no longer do any locking on its own.

As a result of this, I'm no longer making the begin/end/rollback transaction
functions public because you can't start a transaction from within another
transaction.

Of course, doing this means that we aren't guaranteeing writes to a caller of
an ast_db_ function, but that is just the way things go I suppose.

We dropped from being 5x slower than BDB to 4x. Wheeee!