Fixup a race condition in res_fax.c where FAXOPT(gateway)=no will turn off the gateway but the framehook is not destroyed.
this problem happens when a gateway is attempted in the dialplan and
the device is not available i may want to do fax to mail in the server
it will not be allowed.

instead of checking only AST_FAX_TECH_GATEWAY also check gateway_id

Reverts 338904

Fix some white space.