Updated chan_sip to place calls on hold if SDP address in INVITE is ANY
This patch fixes the case where an INVITE is received with c=0.0.0.0 or ::.
In this case, the call should be placed on hold.  Previously, we checked for
the address being null; this patch keeps that behavior but also checks for
the ANY IP addresses.

Review: https://reviewboard.asterisk.org/r/1504/

(closes issue ASTERISK-18086)
Reported by: James Bottomley
Tested by: Matt Jordan