Send FastStart proposal with facility msg when media address changed instead of CloseLC/EmptyTCS if there was no previous H.245 exchange
create rtp instance as fast as we can due to instance is used to get
data for remote bridging

introduce directrtp for remote bridging
introduce earlydirectrtp for early remote bridging (activated by default
if directrtp is on)

Direct rtp support is near of production. Last thing here is to get 
rtp info from the setup packet before LCs are started.